Changing license to Apache, Renaming packages of step-library-selenium
