+++ keywords/java/step-library-kw-selenium/pom.xml
@@ -45,15 +45,5 @@
 			<artifactId>step-library-selenium-3</artifactId>
 			<version>${project.version}</version>
 		</dependency>
-		<dependency>
-			<groupId>org.seleniumhq.selenium</groupId>
-			<artifactId>selenium-java</artifactId>
-			<version>3.141.59</version>
-		</dependency>
-		<dependency>
-			<groupId>net.lightbody.bmp</groupId>
-			<artifactId>browsermob-core</artifactId>
-			<version>2.1.5</version>
-		</dependency>
 	</dependencies>
 </project>
