+++ keywords/java/step-library-selenium-3/src/main/java/ch/exense/step/library/selenium/JSWaiter.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  ******************************************************************************/
-package ch.exense.step.examples.selenium.helper;
+package ch.exense.step.library.selenium;
 
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
